refactor:  퀴즈타입 선택 레이아웃 분리
