+++ src/components/domain/quiz/Start.tsx
@@ -6,7 +6,7 @@ const Start = () => {
         <p>실무 용어 퀴즈를 진행하러 가볼까요?</p>
       </div>
       <div className="mt-10 h-80 bg-gray-800" />
-      <button className="mt-5 px-6 w-full h-14 bg-gradient-to-tr from-gradient-201 to-gradient-202 to-70% rounded-md text-onSurface-300">
+      <button className="mt-52 px-6 w-full h-14 bg-gradient-to-tr from-gradient-201 to-gradient-202 to-70% rounded-md text-onSurface-300">
         시작하기
       </button>
     </div>
